Move mod structure into files
