+++ src/midi/attribute.rs
@@ -1,0 +1,3 @@
+pub fn release_note(message: &[u8]) -> bool {
+  message[2] == 0
+}
